Add Collections Navbar
This update applies website layout formatting. It adds a Navbar to
the left side of the page. The Navbar allows for easy navigation
of collections. It adds a dividing line between sections, and
adds a footer to allow for a bottom dividing line.

This leaves the main Home Page to work as a repository for the
_posts collection, which lists entires specifically by date.